Finish up the 3.75m hab exterior
Work needs to be done to get things set up for baking, but the overall
design of the hab exterior is done.

Hab interior itself is also done, including modeling of structural
volume for mass calculations (minus the toilet area), but volume
calculations need tweaking to reflect the model changes.